+++ formset/formfields/__init__.py
@@ -4,11 +4,14 @@ from django.forms import fields
 
 from formset.formfields.activator import Activator
 from formset.formfields.collection import CollectionField
+from formset.formfields.ranges import DateRangeField, DateTimeRangeField
 from formset.formfields.richtext import RichTextField
 from formset.utils import FileFieldMixin
 from formset.widgets import DualSortableSelector, UploadedFileInput
 
-__all__ = ['Activator', 'FileField', 'RichTextField', 'SortableManyToManyField']
+__all__ = [
+    'Activator', 'DateRangeField', 'DateTimeRangeField', 'FileField', 'RichTextField', 'SortableManyToManyField'
+]
 
 
 class FileField(FileFieldMixin, fields.FileField):